requirements.txt
flask-sqlalchemy==3.0.2
Flask==2.2.2
psycopg2==2.9.5
python-dotenv==0.21.0
pandas==1.5.2
numpy==1.24.0
nltk==3.8.1
scikit-learn==1.2.0
lexicalrichness==0.4.0
